backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.21
sqlalchemy==2.0.35
python-dotenv==1.0.1
pydantic==2.9.2
pyjwt==2.9.0
requests==2.32.3
urllib3==2.2.3
slowapi==0.1.9